feat: Break out Palette v3 Molecules into stories
