+++ src/v2/Apps/Auctions/Components/__tests__/AuctionArtworksRailPlaceholder.jest.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { mount } from "enzyme"
-import { AuctionArtworksRailPlaceholder } from "../AuctionArtworksRail/AuctionArtworksRailPlaceholder"
+import { AuctionArtworksRailPlaceholder } from "../AuctionArtworksRailPlaceholder"
 
 jest.unmock("react-relay")
