+++ src/v2/Components/__stories__/Components/ContentMetadataExample.tsx
@@ -1,0 +1,26 @@
+import { Text } from "@artsy/palette"
+import React from "react"
+
+export const ContentMetadataExample = () => {
+  return (
+    <>
+      <Text variant="xs" textTransform="uppercase" mb={1}>
+        Category Title
+      </Text>
+
+      <Text variant="xl">Section Title or Headline</Text>
+
+      <Text variant="xl" color="black60" mb={1}>
+        Secondary Text
+      </Text>
+
+      <Text variant="sm" mb={4}>
+        Mixtape tumeric edison bulb roof party, chambray street art portland
+        live-edge. Normcore flannel crucifix tattooed. Microdosing hexagon
+        flannel readymade beard hashtag, chillwave woke fixie. Franzen ethical
+        organic, kale chips readymade echo park scenester art party chambray
+        cred raclette.
+      </Text>
+    </>
+  )
+}
