+++ src/v2/Components/__stories__/Components/TombstoneExample.tsx
@@ -1,0 +1,16 @@
+import { Text } from "@artsy/palette"
+import React from "react"
+
+export const TombstoneExample = () => {
+  return (
+    <>
+      <Text variant="md">Artist Name</Text>
+      <Text variant="md" color="black60">
+        Artwork Title, 1900
+      </Text>
+      <Text variant="xs" fontWeight="bold">
+        $100
+      </Text>
+    </>
+  )
+}
